additional utility method to compute payment references for meta pn
