+++ packages/payment-detection/src/index.ts
@@ -16,6 +16,8 @@ import {
   formatAddress,
   getPaymentNetworkExtension,
   getPaymentReference,
+  getPaymentReferencesForMetaPnRequest,
+  flattenRequestByPnId,
   hashReference,
   padAmountForChainlink,
   parseLogArgs,
@@ -61,8 +63,10 @@ export {
   padAmountForChainlink,
   unpadAmountFromChainlink,
   calculateEscrowState,
+  flattenRequestByPnId,
   getPaymentNetworkExtension,
   getPaymentReference,
+  getPaymentReferencesForMetaPnRequest,
   hashReference,
   formatAddress,
 };
