+++ packages/payment-processor/src/payment/encoder-approval.ts
@@ -12,8 +12,10 @@ import {
   hasErc20ApprovalForSwapWithConversion,
   prepareApprovalErc20ForSwapWithConversionToPay,
 } from './swap-conversion-erc20';
-import { getPaymentNetworkExtension } from '@requestnetwork/payment-detection';
-import { flattenRequestByPnId } from './utils';
+import {
+  getPaymentNetworkExtension,
+  flattenRequestByPnId,
+} from '@requestnetwork/payment-detection';
 
 /**
  * For a given request and user, encode an approval transaction if it is needed.
